Move map display methods to separate partial class
